requirements.txt
PySide6
pycryptodome
PyPdf2
